+++ libs/ui/.storybook/main.js
@@ -1,5 +1,8 @@
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
 const rootWebpackConfig = require('../../../.storybook/webpack.config');
+const path = require('path');
+
+const toPath = _path => path.join(process.cwd(), _path);
 
 module.exports = {
   typescript: {
@@ -21,13 +24,30 @@ module.exports = {
     config.resolve.plugins
       ? config.resolve.plugins.push(tsPaths)
       : (config.resolve.plugins = [tsPaths]);
+
+    config.resolve.alias = {
+      ...config.resolve.alias,
+      '@emotion/core': toPath('node_modules/@emotion/react'),
+      '@emotion/styled': toPath('node_modules/@emotion/styled'),
+      '@emotion/styled-base': toPath('node_modules/@emotion/styled'),
+      'emotion-theming': toPath('node_modules/@emotion/react'),
+    };
+
     return config;
   },
+  babel: async options => ({
+    ...options,
+
+    plugins: [
+      '@babel/plugin-proposal-class-properties',
+      '@babel/plugin-proposal-private-methods',
+    ],
+  }),
   stories: ['../src/lib/**/*.stories.@(mdx|jsx|tsx)'],
   addons: [
+    '@storybook/addon-docs',
     '@storybook/addon-essentials',
     '@storybook/addon-notes/register-panel',
-    '@storybook/addon-docs',
     '@storybook/addon-a11y/register',
   ],
 };