+++ libs/ui/src/lib/containers/Card/CardContent/CardContent.style.ts
@@ -1,17 +1,11 @@
 import styled from '@emotion/styled';
-import { css } from '@emotion/core';
+import { css } from '@emotion/react';
 
 // eslint-disable-next-line import/no-unresolved
 import { pxToRem } from '@guiabolsobr/utils';
 
-import { CustomThemeProps } from '../../../../typings/CustomThemeProps';
-
-type CardContentProps = {
-  theme: CustomThemeProps;
-};
-
 export const CardContent = styled.div`
-  ${({ theme }: CardContentProps) => css`
+  ${({ theme }) => css`
     color: ${theme?.colors?.neutral?.dark};
   `};