+++ libs/ui/src/lib/forms/Toggle/Toggle.style.ts
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { css } from '@emotion/core';
+import { css } from '@emotion/react';
 
 import { default as MuiFormControlLabel } from '@material-ui/core/FormControlLabel';
 import { CustomThemeProps } from '../../../typings/CustomThemeProps';