+++ libs/ui/src/lib/general/Theme/Theme.tsx
@@ -1,13 +1,12 @@
 import React from 'react';
 
-import { ThemeProvider, ThemeProviderProps } from 'emotion-theming';
+import { ThemeProvider, ThemeProviderProps } from '@emotion/react';
 
-import { guiaBolso, connect } from '../../../designSystem/tokens';
 import GlobalStyle from '../../../globalStyles';
 
 type ThemeProps = {
   children: React.ReactNode;
-} & ThemeProviderProps<typeof guiaBolso | typeof connect>;
+} & ThemeProviderProps;
 
 const Theme = ({ children, theme }: ThemeProps) => {
   return (