+++ libs/ui/src/lib/notification/infoPanel/InfoPanel.style.ts
@@ -1,13 +1,12 @@
 import styled from '@emotion/styled';
-// eslint-disable-next-line import/no-unresolved
+import { Theme } from '@emotion/react';
 import { pxToRem } from '@guiabolsobr/utils';
-import { CustomThemeProps } from '../../../typings/CustomThemeProps';
 
 import { Text } from '../../display/Text';
 import { FeedbackType } from './InfoPanel';
 
 type InfoPanelTheme = {
-  theme: CustomThemeProps;
+  theme?: Theme;
 };
 
 type ContainerProps = InfoPanelTheme & {