+++ libs/utils/CHANGELOG.md
@@ -7,10 +7,9 @@
 
 ## @guiabolsobr/utils [1.5.0](https://github.com/GuiaBolso/pink-lemonade-web/compare/@guiabolsobr/utils@1.4.1...@guiabolsobr/utils@1.5.0) (2021-06-11)
 
-
 ### Features
 
-* 🎸 add lib emotion-query ([32cb0fe](https://github.com/GuiaBolso/pink-lemonade-web/commit/32cb0fe8aaf619ff78a6f65a88c62c094ca3f8b6))
+- 🎸 add lib emotion-query ([32cb0fe](https://github.com/GuiaBolso/pink-lemonade-web/commit/32cb0fe8aaf619ff78a6f65a88c62c094ca3f8b6))
 
 ### @guiabolsobr/utils [1.4.1](https://github.com/GuiaBolso/pink-lemonade-web/compare/@guiabolsobr/utils@1.4.0...@guiabolsobr/utils@1.4.1) (2021-05-31)